Seperated logic for end game states
