+++ src/logic/moveRules.js
@@ -9,8 +9,8 @@ import {
 } from '../utils/helpers.js';
 import {
     WHITE_PIECES,
-    BLACK_PIECES,
 } from '../constants/pieces.js';
+import { isKingInCheck } from './gameStatus.js';
 
 // ---- Piece move generators ----
 
@@ -290,46 +290,6 @@ function isSquareAttacked(board, row, col, attackerIsWhite) {
   return false;
 }
 
-// Returns the positions of any pieces checking the king of the given color
-function getCheckingPieces(board, color) {
-  const kingPos = findKingPosition(board, color === 'black');
-  if (!kingPos) return null;
-
-  const attackerIsWhite = color === 'black';
-  const attackers = [];
-
-  const isEnemy = (piece) =>
-    attackerIsWhite ? isWhitePiece(piece) : isBlackPiece(piece);
-
-  const getAllEnemyMoves = (r, c, piece) => {
-    if (piece === '♙' || piece === '♟') return getValidPawnMoves(board, r, c, piece, null);
-    if (piece === '♘' || piece === '♞') return getValidKnightMoves(board, r, c, piece);
-    if (piece === '♗' || piece === '♝') return getValidBishopMoves(board, r, c, piece);
-    if (piece === '♖' || piece === '♜') return getValidRookMoves(board, r, c, piece);
-    if (piece === '♕' || piece === '♛') return getValidQueenMoves(board, r, c, piece);
-    if (piece === '♔' || piece === '♚') return [];
-    return [];
-  };
-
-  for (let r = 0; r < 8; r++) {
-    for (let c = 0; c < 8; c++) {
-      const piece = board[r][c];
-      if (!isEnemy(piece)) continue;
-
-      const moves = getAllEnemyMoves(r, c, piece);
-      for (const [mr, mc] of moves) {
-        if (mr === kingPos.row && mc === kingPos.col) {
-          attackers.push({ row: r, col: c });
-          break;
-        }
-      }
-    }
-  }
-
-  if (attackers.length === 0) return null;
-  return { king: kingPos, attackers };
-}
-
 // Helper used for check logic to locate a specific colour king on the board.
 function findKingPosition(board, isWhite) {
   const kingSymbol = isWhite ? '♚' : '♔'; // enemy king
@@ -341,19 +301,6 @@ function findKingPosition(board, isWhite) {
     }
   }
   return null;
-}
-
-// Wrapper around isSquareAttacked that first finds the king of the given color
-// and determines if any opposing piece can attack that square.
-function isKingInCheck(board, color) {
-  const kingPos = findKingPosition(board, color === 'black');
-  if (!kingPos) return false;
-  return isSquareAttacked(
-    board,
-    kingPos.row,
-    kingPos.col,
-    color === 'black'
-  );
 }
 
 // Helper used when checking move legality. It plays a move on a cloned board
@@ -411,36 +358,6 @@ function filterLegalMoves(moves, board, fromRow, fromCol, piece, enPassantTarget
     if (!isKingInCheck(newBoard, color)) legal.push(move);
   }
   return legal;
-}
-
-// Iterates all pieces of a given color and checks if at least one legal move
-// exists. Used for stalemate/checkmate detection.
-function hasAnyLegalMoves(board, color, kingState, enPassantTarget, castlingRights) {
-  for (let r = 0; r < 8; r++) {
-    for (let c = 0; c < 8; c++) {
-      const piece = board[r][c];
-      if (color === 'white' && !WHITE_PIECES.includes(piece)) continue;
-      if (color === 'black' && !BLACK_PIECES.includes(piece)) continue;
-
-      let moves = [];
-      if (piece === '♙' || piece === '♟')
-        moves = getValidPawnMoves(board, r, c, piece, enPassantTarget);
-      else if (piece === '♖' || piece === '♜')
-        moves = getValidRookMoves(board, r, c, piece);
-      else if (piece === '♕' || piece === '♛')
-        moves = getValidQueenMoves(board, r, c, piece);
-      else if (piece === '♘' || piece === '♞')
-        moves = getValidKnightMoves(board, r, c, piece);
-      else if (piece === '♗' || piece === '♝')
-        moves = getValidBishopMoves(board, r, c, piece);
-      else if (piece === '♔' || piece === '♚')
-        moves = getValidKingMoves(board, r, c, piece, kingState[color], castlingRights);
-
-      const legal = filterLegalMoves(moves, board, r, c, piece, enPassantTarget);
-      if (legal.length > 0) return true;
-    }
-  }
-  return false;
 }
 
 // Main king move generator. Handles normal one-square movement while ensuring
@@ -563,13 +480,9 @@ export {
     getValidBishopMoves,
     getValidKingMoves,
     isSquareAttacked,
-    getCheckingPieces,
     findKingPosition,
-    isKingInCheck,
     simulateMove,
     filterLegalMoves,
-    hasAnyLegalMoves,
     performSummon,
     performPromotion,
   };
-
