+++ src/utils/helpers.js
@@ -17,54 +17,6 @@ const cloneBoard = (board) => board.map((r) => [...r]);
 // Utility to deep clone objects (used for history snapshots)
 const deepClone = (obj) => JSON.parse(JSON.stringify(obj));
 
-// Serialize board position plus turn/castling/en passant to detect repetitions
-function boardKey(board, turn, castlingRights, enPassantTarget) {
-  const map = {
-    '♙': 'P',
-    '♟': 'p',
-    '♘': 'N',
-    '♞': 'n',
-    '♖': 'R',
-    '♜': 'r',
-    '♗': 'B',
-    '♝': 'b',
-    '♕': 'Q',
-    '♛': 'q',
-    '♔': 'K',
-    '♚': 'k',
-  };
-  const rows = board.map((row) => {
-    let str = '';
-    let empty = 0;
-    for (const piece of row) {
-      if (piece === '') {
-        empty += 1;
-      } else {
-        if (empty > 0) {
-          str += empty;
-          empty = 0;
-        }
-        str += map[piece] || piece;
-      }
-    }
-    if (empty > 0) str += empty;
-    return str;
-  });
-  const boardStr = rows.join('/');
-  const rights = [
-    castlingRights.white.kingSide ? 'K' : '',
-    castlingRights.white.queenSide ? 'Q' : '',
-    castlingRights.black.kingSide ? 'k' : '',
-    castlingRights.black.queenSide ? 'q' : '',
-  ]
-    .join('') || '-';
-  const ep = enPassantTarget
-    ? String.fromCharCode(97 + enPassantTarget.col) + (8 - enPassantTarget.row)
-    : '-';
-  const turnChar = turn === 'white' ? 'w' : 'b';
-  return `${boardStr} ${turnChar} ${rights} ${ep}`;
-}
-
 export {
   cloneBoard,
   deepClone,
@@ -72,5 +24,4 @@ export {
   isBlackPiece,
   isEnemyPiece,
   isSameTeam,
-  boardKey,
 };
